preparing pipeline and frame graph for render to texture

+++ game/cooker/src/cook_pipeline.rs
@@ -1,6 +1,5 @@
 use crate::{cook_shader, AssetNaming, Context, CookingError, Dependency};
-use shine_game::assets::{AssetId, PipelineDescriptor, Url};
-use shine_game::render::MAX_UNIFORM_GROUP_COUNT;
+use shine_game::assets::{AssetId, PipelineDescriptor, Url, MAX_UNIFORM_GROUP_COUNT};
 
 async fn find_pipeline_etag(context: &Context, pipeline_url: &Url) -> Result<String, CookingError> {
     Ok(context.source_io.download_etag(&pipeline_url).await?)
@@ -33,7 +32,7 @@ pub async fn cook_pipeline(
     let mut dependencies = Vec::new();
 
     for i in 0..MAX_UNIFORM_GROUP_COUNT {
-        let layout = pipeline.get_uniform_layout(i)?;
+        let layout = pipeline.get_uniform_layout(i as u32)?;
         log::trace!(
             "[{}] Uniform group({}) layout:\n{:#?}",
             pipeline_url.as_str(),
